requirements.txt
httpx>=0.27
tomli>=2.0; python_version < "3.11"
packaging>=24.0
pyyaml>=6.0
